+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added support for formatting dynamic messages using the protobuf [text format](https://developers.google.com/protocol-buffers/docs/text-format-spec), through the new [`Display`](https://docs.rs/prost-reflect/latest/prost_reflect/struct.DynamicMessage.html#impl-Display) implementation for `DynamicMessage`.
+- Added new methods for setting dynamic message fields without panicking: [`try_set_field`](https://docs.rs/prost-reflect/latest/prost_reflect/struct.DynamicMessage.html#method.try_set_field), [`try_set_field_by_number`](https://docs.rs/prost-reflect/latest/prost_reflect/struct.DynamicMessage.html#method.try_set_field_by_number) and [`try_set_field_by_name`](https://docs.rs/prost-reflect/latest/prost_reflect/struct.DynamicMessage.html#method.try_set_field_by_name)
 
 ## [0.9.1] - 2022-08-01